rename 'sanctuaryEnvirnoment' to 'sanctuaryEnvironment' rename file to remove typo update references to new file name
