+++ lib/ftp.js
@@ -1,9 +1,11 @@
 const $ = require('sanctuary-def');
 const S = require('sanctuary');
 const { FutureType } = require('fluture-sanctuary-types')
-const { def, FtpConnectionConfig } = require('../lib/sanctuaryEnvirnoment');
-const {Future, reject, resolve} = require('fluture');
+const { def, FtpConnectionConfig } = require('./sanctuaryEnvironment');
+const {Future} = require('fluture');
+
 // validate incoming FtpConnection
+
 // sendFile :: String -> FtpConnectionConfig -> Future
 const sendFile =
     def ('sendFile')
