+++ lib/sanctuaryEnvironment.js
